+++ R/rdeltanci.R
@@ -9,9 +9,12 @@ NULL
 #' parameters mu and sigma. Default parameters correspond to values inferred in
 #' the control and biodiversity plots of Paracou.
 #'
-#' @param n int. Number of observations.
-#' @param p double. Probability to be negative used in the Bernoulli
-#'   distribution.
+#' @param NCI double. Vector of previous NCI (size of the vector determines
+#'   number of generated deltaNCI).
+#' @param alpha double. Intercept for the Bernoulli distribution determining the
+#'   risk to have a negative deltaNCI.
+#' @param beta double. Slope of previous NCI for the Bernoulli distribution
+#'   determining the risk to have a negative deltaNCI.
 #' @param mu double. mu parameter for the lognormal distribution used for
 #'   positive deltaNCI.
 #' @param sigma double. sigma parameter for the lognormal distribution used for
@@ -26,15 +29,16 @@ NULL
 #' rdeltanci(1)
 #' 
 rdeltanci <- function(
-  n = 1, # number of samples
-  p = 0.271, # probability to be positive
-  mu = 0.749, # mu of lognormal distributon for positive deltanci
-  sigma = 2.651, # sigma of lognormal distributon for positive deltanci
-  lambda = 0.31 # lamnda of exponential distributon for negative deltanci
+  NCI = rnorm(1, mean = 122.73, sd = 25.806),
+  alpha = -1.32,
+  beta = 0.003,
+  mu = 0.749,
+  sigma = 2.651,
+  lambda = 0.31
 ){
-  samples <- rep(0, n)
-  P <- rbinom(n, 1, 1 - p) == 1
-  samples[P] <- rlnorm(sum(P), meanlog = log(mu), sdlog = log(sigma))
-  samples[!P] <- -rexp(sum(!P), rate = lambda)
+  samples <- rep(0, length(NCI))
+  P <- sapply(invlogit(alpha + beta*NCI), function(p) rbinom(1,1,p)) == 1
+  samples[!P] <- rlnorm(sum(!P), meanlog = log(mu), sdlog = log(sigma))
+  samples[P] <- -rexp(sum(P), rate = lambda)
   return(samples)
 }