+++ R/simulator.R
@@ -9,23 +9,26 @@ NULL
 #'
 #' @param grid int.  Number of cells per side of the matrix.
 #' @param Nt int. Number of time steps.
+#' @param timestep int. Time-step length in years.
 #' @param topography char. Topography generator between 'sinusoidal',
 #'   'squarediamond' or 'paracou'.
 #' @param Elim double. Environmental matrix extrme (absolute value).
 #' @param amplitude double. Amplitude of the sinusoidal functional.
-#' @param rudgeness double. Rugedness parameter.
+#' @param ruggedness double. Ruggedness parameter.
 #' @param muNCI double. mu parameter for the normal distribution used for NCI.
 #' @param sigmaNCI double. sigma parameter for the normal distribution used for
 #'   NCI.
-#' @param p double. Probability to be negative used in the Bernoulli
-#'   distribution.
+#' @param alpha double. Intercept for the Bernoulli distribution determining the
+#'   risk to have a negative deltaNCI.
+#' @param beta double. Slope of previous NCI for the Bernoulli distribution
+#'   determining the risk to have a negative deltaNCI.
 #' @param mu double. mu parameter for the lognormal distribution used for
 #'   positive deltaNCI.
 #' @param sigma double. sigma parameter for the lognormal distribution used for
 #'   positive deltaNCI.
 #' @param lambda double. lambda parameter for the exponential distribution used
 #'   for negative deltaNCI.
-#' @param d int. Spatial autoccorelation size in number of cells.
+#' @param d int. Spatial auto-correlation size in number of cells.
 #' @param plot int. Plot number between 1 and 15 (a cell size is 3x3m).
 #' @param sigmaGtopo double. Variance of genetic values with topography.
 #' @param sigmaZtopo double. Plasticity of phenotypes with topography.
@@ -33,8 +36,9 @@ NULL
 #' @param sigmaZnci double. Plasticity of phenotypes with NCI.
 #' @param Pdeath double. Background mortality probability.
 #' @param Ns int. Number of seedlings per cell.
-#' @param Rdispersal int. Dispersal radius in cells.
-#' @param determinist bool. Deterministic or probabilistic vaibility.
+#' @param Rpollination int. Pollination radius in cells (father to mother).
+#' @param Rdispersion int. Dispersal radius in cells (mother to seedling).
+#' @param determinist bool. Deterministic or probabilistic viability.
 #' @param verbose bool. Should the function print statuses.
 #'
 #' @return A data frame.
@@ -48,25 +52,28 @@ NULL
 simulator <- function(
   grid = 20,
   Nt = 50,
+  timestep = 30, # time-step length in years
   topography = "sinusoidal",
   Elim = 5,
   amplitude = 1,
-  rudgeness = 1,
+  ruggedness = 1,
   plot = 1,
-  muNCI = 124, # mu of normal distributon for nci
-  sigmaNCI = 26, # sigma of normal distributon nci
-  p = 0.271, # probability to be positive
-  mu = 0.749, # mu of lognormal distributon for positive deltanci
-  sigma = 2.651, # sigma of lognormal distributon for positive deltanci
-  lambda = 0.31, # lamnda of exponential distributon for negative deltanci
-  d = 3, # spatial autocorrelation size (3*3m)
+  muNCI = 124, # mu of normal distribution for NCI
+  sigmaNCI = 26, # sigma of normal distribution NCI
+  alpha = -1.32,
+  beta = 0.003,
+  mu = 0.749, # mu of lognormal distribution for positive deltaNCI
+  sigma = 2.651, # sigma of lognormal distribution for positive deltaNCI
+  lambda = 0.31, # lambda of exponential distribution for negative deltaNCI
+  d = 3, # spatial auto-correlation size (3*3m)
   sigmaGtopo = 1,
   sigmaZtopo = 1,
   sigmaGnci = 26,
   sigmaZnci = 26,
   Pdeath = 0.01325548,
   Ns = 4,
-  Rdispersal = 1,
+  Rpollination = 1,
+  Rdispersion = 1,
   determinist = TRUE,
   verbose = TRUE
 ){
@@ -79,7 +86,7 @@ simulator <- function(
   if(topography == 'sinusoidal')
     Topo <- sinusoidalTopography(grid = grid, Elim = Elim, amplitude = amplitude)
   if(topography == 'squarediamond')
-    Topo <- squareDiamondTopography(grid = grid,  rudgeness = rudgeness)
+    Topo <- squareDiamondTopography(grid = grid,  ruggedness = ruggedness)
   if(topography == 'paracou')
     Topo <- paracouTopography(grid = grid, plot = plot, Elim = Elim)
   
@@ -88,9 +95,11 @@ simulator <- function(
   NCI <- generateNCIsim(
     grid = grid, 
     Nt = Nt, 
+    timestep = timestep,
     muNCI = muNCI, 
     sigmaNCI = sigmaNCI, 
-    p = p, 
+    alpha = alpha,
+    beta = beta,
     mu = mu,
     sigma = sigma,
     lambda = lambda,
@@ -103,13 +112,15 @@ simulator <- function(
     NCI = NCI,
     grid = grid, 
     Nt = Nt, 
+    timestep = timestep,
     sigmaGtopo = sigmaGtopo,
     sigmaZtopo = sigmaZtopo,
     sigmaGnci = sigmaGnci,
     sigmaZnci = sigmaZnci,
     Pdeath = Pdeath, 
     Ns = Ns, 
-    Rdispersal = Rdispersal, 
+    Rpollination = Rpollination,
+    Rdispersion = Rdispersion, 
     determinist =  determinist
   )
   coords <- data.frame(